ref: revert unnecessary move of asset entry dto
